add repository

+++ src/main/java/com/pm/repository/RoleRepository.java
@@ -1,8 +1,7 @@
 package com.pm.repository;
 
 import com.pm.entity.RoleEntity;
-import com.pm.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleRepository extends JpaRepository<RoleEntity,Long> {
+public interface RoleRepository extends JpaRepository<RoleEntity,Integer> {
 }
